LAYOUT SCHEMATIC / MOCKUP
Yo, attached is a detailed schematic of how the header and blog page
should layout. This is gonna be the frame for the site. Notice the
lower section will be the main content area  for each section of the
site. I'm going to post a discussion to share these and explain
further, but this should be some solid shit to get going. I'll be
finishing up the looks for all the other pages over the next day or so.
Sorry for all the changes since last time and for taking a while to get
this shit together. I was kinda stuck for a minute, and being hella
busy at work doesnt help. I'll hit you up tomorrow.